Added ActivityUseButton component.
Added View context menu option to locked compendia and unowned content.

Factored out context menu handling into distinct files per type of context menu, to provide better scaling.

Took over activity context menu management, finally.

Increased the amount of data put into item context linkedUses, to allow for updating linked uses in the attribute pins.

Added mode forwarding based on Tidy's sheet lock state, to mirror the default sheet's approach. This is for users who use tidy actor sheets and default item sheets.

Added support for linked uses in attr item pins.

Fixed context menu issue with favorite effects.

Added error boundary to individual pins, with logging.
