+++ src/constants.ts
@@ -48,6 +48,7 @@ export const CONSTANTS = {
    * Secondary button pressed, usually the right button
    */
   MOUSE_BUTTON_SECONDARY: 2,
+  CONTEXT_MENU_TYPE_ACTIVITIES: 'activities',
   CONTEXT_MENU_TYPE_EFFECTS: 'effects',
   CONTEXT_MENU_TYPE_FACILITY_OCCUPANTS: 'facility-occupants',
   CONTEXT_MENU_TYPE_GROUP_MEMBER: 'group-member',
